backend/requirements.txt
flask
numpy
scikit-image
scikit-learn
opencv-python
pillow
matplotlib
joblib
tiffile
imagecodecs
rasterio